third_party/WebKit/LayoutTests/http/tests/devtools/websocket/websocket-frame-error-expected.txt

CONSOLE ERROR: line 8: WebSocket connection to 'ws://localhost:8000/does_not_exist' failed: Error during WebSocket handshake: Unexpected response code: 404
PASS successfullyParsed is true

TEST COMPLETE
Closed.
Tests that WebSocketFrames errors are visible to Web Inspector.

1-error: Error during WebSocket handshake: Unexpected response code: 404
